+++ src/server/graphql/blockchain/blockchain.ts
@@ -11,8 +11,14 @@ import {
     resolve_account_transactions,
     resolve_key_blocks,
     resolve_blockchain_blocks,
+    resolve_account_messages,
     resolve_blockchain_transactions,
+    resolve_block,
+    resolve_transaction,
+    resolve_message,
+    resolve_block_by_seq_no,
 } from "./fetchers";
+
 
 function parseMasterSeqNo(chain_order: string) {
     const length = parseInt(chain_order[0], 16) + 1;
@@ -134,6 +140,26 @@ export const resolvers: Resolvers<QRequestContext> = {
                 address: args.address
             }
         },
+        block: async (_parent, args, context, info) => {
+            return context.trace("blockchain-block", async traceSpan => {
+                return resolve_block(args.hash, context, info, traceSpan);
+            });
+        },
+        block_by_seq_no: async (_parent, args, context, info) => {
+            return context.trace("blockchain-block-by-seq-no", async traceSpan => {
+                return resolve_block_by_seq_no(args, context, info, traceSpan);
+            });
+        },
+        transaction: async (_parent, args, context, info) => {
+            return context.trace("blockchain-transaction", async traceSpan => {
+                return resolve_transaction(args.hash, context, info, traceSpan);
+            });
+        },
+        message: async (_parent, args, context, info) => {
+            return context.trace("blockchain-message", async traceSpan => {
+                return resolve_message(args.hash, context, info, traceSpan);
+            });
+        },
         master_seq_no_range: (_parent, args, context) => {
             return context.trace("blockchain-master_seq_no_range", async traceSpan => {
                 return await resolve_maser_seq_no_range(args, context, traceSpan);
@@ -175,15 +201,20 @@ export const resolvers: Resolvers<QRequestContext> = {
     },
     BlockchainAccountQuery: {
         info: async (parent, _args, context, info) => {
-            return context.trace("account-info", async traceSpan => {
+            return context.trace("blockchain-account-info", async traceSpan => {
                 return resolve_account(parent.address, context, info, traceSpan);
             });
         }, 
         transactions: async (parent, args, context, info) => {
-            return context.trace("account-transactions", async traceSpan => {
+            return context.trace("blockchain-account-transactions", async traceSpan => {
                 return await resolve_account_transactions(parent.address, args, context, info, traceSpan);
             });
         },
+        messages: async (parent, args, context, info) => {
+            return context.trace("blockchain-account-messages", async traceSpan => {
+                return await resolve_account_messages(parent, args, context, info, traceSpan);
+            });
+        }
     },
     Node: {
         __resolveType: (parent) => {