+++ src/server/graphql/blockchain/fetchers/blocks.ts
@@ -18,9 +18,86 @@ import {
     BlockchainBlock,
     BlockchainBlocksConnection,
     BlockchainQueryBlocksArgs,
+    BlockchainQueryBlock_By_Seq_NoArgs,
     BlockchainQueryKey_BlocksArgs,
 } from "../resolvers-types-generated";
 
+async function fetch_block(
+    filterBuilder: (params: QParams) => String,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+    maxJoinDepth = 1,
+    shards?: Set<string>,
+) {
+    const selectionSet = info.fieldNodes[0].selectionSet;
+    const returnExpression = config.blocks.buildReturnExpression(
+        selectionSet,
+        context,
+        maxJoinDepth,
+        "doc"
+    );
+
+    // query
+    const params = new QParams();
+    const query =
+        "FOR doc IN blocks " +
+        `FILTER ${filterBuilder(params)} ` +
+        `RETURN ${returnExpression}`;
+    const queryResult = await context.services.data.query(
+        required(context.services.data.blocks.provider),
+        {
+            text: query,
+            vars: params.values,
+            orderBy: [],
+            request: context,
+            traceSpan,
+            shards,
+        },
+    ) as BlockchainBlock[];
+
+    await config.blocks.fetchJoins(
+        queryResult,
+        selectionSet,
+        context,
+        traceSpan,
+        maxJoinDepth,
+    );
+
+    return queryResult[0];
+}
+
+export async function resolve_block(
+    hash: String,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+) {
+    return fetch_block(
+        (params) => `doc._key == @${params.add(hash)}`,
+        context,
+        info,
+        traceSpan,
+        // TODO: shard
+    );
+}
+
+export async function resolve_block_by_seq_no(
+    args: BlockchainQueryBlock_By_Seq_NoArgs,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+) {
+    return fetch_block(
+        (params) => 
+            `doc.workchain_id == @${params.add(args.workchain)} AND ` +
+            `doc.shard == @${params.add(args.thread)} AND ` +
+            `doc.seq_no == @${params.add(args.seq_no)}`,
+        context,
+        info,
+        traceSpan,
+    );
+}
 
 export async function resolve_key_blocks(
     args: BlockchainQueryKey_BlocksArgs,
@@ -76,6 +153,7 @@ export async function resolve_key_blocks(
         queryResult,
         limit,
         direction,
+        "chain_order",
     ) as BlockchainBlocksConnection;
 }
 
@@ -121,7 +199,6 @@ export async function resolve_blockchain_blocks(
         "doc",
     );
 
-
     // query
     const query = `
         FOR doc IN blocks
@@ -150,5 +227,6 @@ export async function resolve_blockchain_blocks(
         queryResult,
         limit,
         direction,
+        "chain_order",
     ) as BlockchainBlocksConnection;
 }