+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.49.0] - 2022-03-29
+
+### New
+
+- New functions:
+    - `blockchain { account { messages }}` to fetch account messages
+    - `blockchain { block }` to fetch block by hash
+    - `blockchain { block_by_seq_no }` to fetch block by (workchain, thread, seq_no) triplet
+    - `blockchain { transaction }` to fetch transaction by hash
+    - `blockchain { message }` to fetch message by hash
+
+### Fixed
+
+- Fixed `created_at_string` and `status_name` fields resolvers in BlockchainMessage
+- Messages subscriptions now ignore patch records
+
+
 ## [0.48.1] - 2022-03-16
 
 ### Fix